requirements.txt
Flask==3.1.0
mysql-connector-python==9.2.0
python-dotenv==1.0.1
Flask-cors==5.0.0
bcrypt==4.2.1
PyJWT==2.8.0
scikit-learn==1.6.1
numpy==2.2.3
pandas==2.2.3
